used coding assistant to restructure project
